# openapi.yml
openapi: '3.0.3'
info:
  title: API Users
  version: '1.0'
servers:
  - url: http://localhost:3000

components:
  schemas:
    user:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        age:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string

paths:
  /users:
    get:
      description: 'Endpoint untuk mendapatkan semua data user'
      responses:
        '200':
          description: Berhasil
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/user'
        '500':
          description: Server Gagal
        '403':
          description: Tidak diizinkan

    post:
      description: 'Endpoint untuk menambahkan pengguna baru'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                age:
                  type: integer
      responses:
        '201':
          description: Pengguna berhasil ditambahkan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user'
        '400':
          description: Request body tidak lengkap
        '500':
          description: Server Gagal

  /users/{id}:
    get:
      description: 'Endpoint untuk mendapatkan data user berdasarkan ID'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Berhasil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user'
        '404':
          description: Pengguna tidak ditemukan
        '500':
          description: Server Gagal

    put:
      description: 'Endpoint untuk mengubah data user berdasarkan ID'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                age:
                  type: integer
      responses:
        '200':
          description: Berhasil diperbarui
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user'
        '400':
          description: Request body tidak lengkap
        '404':
          description: Pengguna tidak ditemukan
        '500':
          description: Server Gagal

    delete:
      description: 'Endpoint untuk menghapus data user berdasarkan ID'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Pengguna berhasil dihapus
        '404':
          description: Pengguna tidak ditemukan
        '500':
          description: Server Gagal